[Strip] make background solid black for maximum contrast
In a controlled, well-lit, desktop work environment the non-black is
just fine and looks real nice, but in these work-from-home times,
working on a laptop is more common and the current contrast ratio just
didn’t cut it. This was especially apparent in brightly-lit, outdoors
conditions or dark, low-screen-brightness situations. The latter was
particularly annoying because a comfortable brightness for reading code
was too bright when switching to an application with a white background
(like a browser). This hopefully balances things out.